+++ src/icons/icon/index.ts
@@ -1,4 +1,4 @@
-export { default as ArrowIcon } from './ArrowIcon';
+export { default as ArrowRightIcon } from './ArrowRightIcon';
 export { default as CalendarIcon } from './CalendarIcon';
 export { default as MarkerIcon } from './MarkerIcon';
 export { default as CheckboxIcon } from './CheckboxIcon';
