+++ src/pages/Reservation/Index/Index.module.scss
@@ -7,19 +7,15 @@
   line-height: 22px; /* 146.667% */
 }
 
-.DividerWrap {
-  margin: 26px 0;
-}
-
 .TimeWrap {
-  display: flex;
   justify-content: space-between;
   text-align: center;
-  flex-wrap: wrap;
+  display: grid;
+  grid-template-columns: repeat(5, minmax(0, 1fr));
+  gap: 8px;
   margin-top: 16px;
   > .TimeButton {
     padding: 12px 14px;
-    flex-basis: calc(20% - 10px); /* 한 행에 최대 5개의 요소가 들어가도록 설정 */
     margin-bottom: 10px; /* 요소들 간의 세로 간격 */
   }
 }
